+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.3.25 (2023-08-18)
+
+### Changed
+
+- Improve Grafana Alerting integration by @Ferril @teodosii ([#2742](https://github.com/grafana/oncall/pull/2742))
+- Fixed UTC conversion for escalation chain step of timerange
+  ([#2781](https://github.com/grafana/oncall/issues/2781))
+
+### Fixed
+
+- Check for possible split events in range when resolving schedule ([#2828](https://github.com/grafana/oncall/pull/2828))
+
 ## v1.3.24 (2023-08-17)
 
 ### Added